d4d_button: enable swapping active bitmap on press
The button may not have D4D_OBJECT_F_TABSTOP enabled and so will never
receive focus. The enables the button to swap the active bitmap depending
on whether the button is pressed or released.

The can be useful when the interface does not have a touchscreen and uses
a hardware keypad only.

Signed-off-by: Spencer Oliver <spen@spen-soft.co.uk>